Rename package to quppulse (move files)
