+++ qupulse/hardware/awgs/base.py
