+++ qupulse/pulses/multi_channel_pulse_template.py
